+++ app/routes/data-api/user.js
@@ -14,8 +14,14 @@ router.get('/projectlist', function(req, res, next) {
     var user = req.session.user;
     var type = req.query.type;
     var includeLocation = req.query.include_location === 'true';
+    let allAccessibleProjects = req.query.allAccessibleProjects === 'true';
     if ((user.isAnonymousGuest || user.isSuper !== 1) && !type) {
-        model.users.projectList(req.session.user.id, function(err, rows) {
+        model.users.projectList({
+            user_id: req.session.user.id,
+            allAccessibleProjects: allAccessibleProjects
+            ...req.query.q && { q: req.query.q },
+            ...req.query.featured && { featured: req.query.featured }
+        }, function(err, rows) {
             if(err) return next(err);
             res.json(rows);
         });
@@ -23,7 +29,10 @@ router.get('/projectlist', function(req, res, next) {
     else {
         model.projects.find({
             ...type === 'my' && { user_id: user.id },
-            ...includeLocation && { include_location: true }
+            ...includeLocation && { include_location: true },
+            allAccessibleProjects: allAccessibleProjects
+            ...req.query.q && { q: req.query.q },
+            ...req.query.featured && type !== 'my' && { featured: req.query.featured }
         }, function(err, rows) {
             if(err) return next(err);
             res.json(rows);